+++ easypost/models/shipment.py
@@ -1,0 +1,16 @@
+from typing import (
+    List,
+    Optional,
+)
+
+from easypost.easypost_object import EasyPostObject
+from easypost.models.rate import Rate
+from easypost.util import get_lowest_object_rate
+
+
+class Shipment(EasyPostObject):
+    def lowest_rate(self, carriers: Optional[List[str]] = None, services: Optional[List[str]] = None) -> Rate:
+        """Get the lowest rate of this shipment."""
+        lowest_rate = get_lowest_object_rate(self, carriers, services)
+
+        return lowest_rate