+++ easypost/models/webhook.py
@@ -1,0 +1,5 @@
+from easypost.easypost_object import EasyPostObject
+
+
+class Webhook(EasyPostObject):
+    pass