+++ easypost/services/order_service.py
@@ -32,8 +32,8 @@ class OrderService(BaseService):
         insurances: Dict[str, Any],
         page_size: int,
         optional_params: Optional[Dict[str, Any]] = None,
-    ) -> List[Any]:
-        """Retrieve the next page of the List Order response."""
+    ) -> List[Order]:
+        """Retrieve the next page of the list Order response."""
         return self._get_next_page_resources(self._model_class, insurances, page_size, optional_params)
 
     def get_rates(self, id: str) -> Order: