+++ tests/test_scan_form.py
@@ -5,30 +5,30 @@ from easypost.error import Error
 
 
 @pytest.mark.vcr()
-def test_scanform_create(one_call_buy_shipment):
-    shipment = easypost.Shipment.create(**one_call_buy_shipment)
+def test_scanform_create(one_call_buy_shipment, test_client):
+    shipment = easypost.Shipment.create(**one_call_buy_shipment)  # TODO: Use new syntax
 
-    scanform = easypost.ScanForm.create(shipment=[shipment])
+    scanform = test_client.scan_form.create(shipment=[shipment])
 
     assert isinstance(scanform, easypost.ScanForm)
     assert str.startswith(scanform.id, "sf_")
 
 
 @pytest.mark.vcr()
-def test_scanform_retrieve(one_call_buy_shipment):
-    shipment = easypost.Shipment.create(**one_call_buy_shipment)
+def test_scanform_retrieve(one_call_buy_shipment, test_client):
+    shipment = easypost.Shipment.create(**one_call_buy_shipment)  # TODO: Use new syntax
 
-    scanform = easypost.ScanForm.create(shipment=[shipment])
+    scanform = test_client.scan_form.create(shipment=[shipment])
 
-    retrieved_scanform = easypost.ScanForm.retrieve(scanform.id)
+    retrieved_scanform = test_client.scan_form.retrieve(scanform.id)
 
     assert isinstance(retrieved_scanform, easypost.ScanForm)
     assert retrieved_scanform == scanform
 
 
 @pytest.mark.vcr()
-def test_scanform_all(page_size):
-    scanforms = easypost.ScanForm.all(page_size=page_size)
+def test_scanform_all(page_size, test_client):
+    scanforms = test_client.scan_form.all(page_size=page_size)
 
     scanforms_array = scanforms["scan_forms"]
 
@@ -38,10 +38,10 @@ def test_scanform_all(page_size):
 
 
 @pytest.mark.vcr()
-def test_scanform_get_next_page(page_size):
+def test_scanform_get_next_page(page_size, test_client):
     try:
-        scanforms = easypost.ScanForm.all(page_size=page_size)
-        next_page = easypost.ScanForm.get_next_page(collection=scanforms, page_size=page_size)
+        scanforms = test_client.scan_form.all(page_size=page_size)
+        next_page = test_client.scan_form.get_next_page(scan_forms=scanforms, page_size=page_size)
 
         first_id_of_first_page = scanforms["scan_forms"][0].id
         first_id_of_second_page = next_page["scan_forms"][0].id