+++ tests/test_tracker.py
@@ -5,8 +5,8 @@ from easypost.error import Error
 
 
 @pytest.mark.vcr()
-def test_tracker_create():
-    tracker = easypost.Tracker.create(tracking_code="EZ1000000001")
+def test_tracker_create(test_client):
+    tracker = test_client.tracker.create(tracking_code="EZ1000000001")
 
     assert isinstance(tracker, easypost.Tracker)
     assert str.startswith(tracker.id, "trk_")
@@ -14,10 +14,10 @@ def test_tracker_create():
 
 
 @pytest.mark.vcr()
-def test_tracker_retrieve():
-    tracker = easypost.Tracker.create(tracking_code="EZ1000000001")
+def test_tracker_retrieve(test_client):
+    tracker = test_client.tracker.create(tracking_code="EZ1000000001")
 
-    retrieved_tracker = easypost.Tracker.retrieve(tracker.id)
+    retrieved_tracker = test_client.tracker.retrieve(tracker.id)
 
     assert isinstance(retrieved_tracker, easypost.Tracker)
     # status changes between creation and retrieval, so we can't compare the whole object
@@ -25,8 +25,8 @@ def test_tracker_retrieve():
 
 
 @pytest.mark.vcr()
-def test_tracker_all(page_size):
-    trackers = easypost.Tracker.all(page_size=page_size)
+def test_tracker_all(page_size, test_client):
+    trackers = test_client.tracker.all(page_size=page_size)
 
     trackers_array = trackers["trackers"]
 
@@ -36,10 +36,10 @@ def test_tracker_all(page_size):
 
 
 @pytest.mark.vcr()
-def test_tracker_get_next_page(page_size):
+def test_tracker_get_next_page(page_size, test_client):
     try:
-        trackers = easypost.Tracker.all(page_size=page_size)
-        next_page = easypost.Tracker.get_next_page(trackers=trackers, page_size=page_size)
+        trackers = test_client.tracker.all(page_size=page_size)
+        next_page = test_client.tracker.get_next_page(trackers=trackers, page_size=page_size)
 
         first_id_of_first_page = trackers["trackers"][0].id
         first_id_of_second_page = next_page["trackers"][0].id
@@ -51,10 +51,10 @@ def test_tracker_get_next_page(page_size):
 
 
 @pytest.mark.vcr()
-def test_tracker_create_list():
+def test_tracker_create_list(test_client):
     """Tests that we can create a list of trackers in bulk."""
     try:
-        easypost.Tracker.create_list(
+        test_client.tracker.create_list(
             {
                 "0": {"tracking_code": "EZ1000000001"},
                 "1": {"tracking_code": "EZ1000000002"},