+++ CHANGELOG.md
@@ -6,8 +6,11 @@
   - Services and models
 - Beta namespace changed from `easypost.beta.x` to `client.beta_x`
 - Empty API response functions return `None` instead of `True`
-- References to `Referral` are now `ReferralCustomer` to match the API and docs
+- References to `Referral` are now `ReferralCustomer` and `referral_customer` to match the API and docs
+- References to `Smartrate` are now `SmartRate` and `smart_rate` to match the API and docs
+- References to `Scanform` are now `ScanForm` and `scan_form`
 - `primary_or_secondary` paramater name for billing functions is now called `priority` to match the API and docs
+- The `update_email` function of the `referral_customer` service had the parameter order switched so `id` (previously called `user_id`) is first which matches the rest of the library
 - Dropped Python 3.6 support
 - Bumps all dependencies